+++ component/pprof/scrape/scrape_loop_test.go
@@ -71,7 +71,7 @@ func TestScrapePool(t *testing.T) {
 				NewTarget(
 					labels.FromStrings("instance", "localhost:8080", "foo", "bar", model.AddressLabel, "localhost:8080", model.MetricNameLabel, pprofProcessCPU, model.SchemeLabel, "http", ProfilePath, "/debug/pprof/profile"),
 					labels.FromStrings("foo", "bar", model.AddressLabel, "localhost:8080", model.MetricNameLabel, pprofProcessCPU, model.SchemeLabel, "http", ProfilePath, "/debug/pprof/profile"),
-					url.Values{"seconds": []string{"14"}},
+					url.Values{"seconds": []string{"17"}},
 				),
 				NewTarget(
 					labels.FromStrings("instance", "localhost:9090", "foo", "bar", model.AddressLabel, "localhost:9090", model.MetricNameLabel, pprofMutex, model.SchemeLabel, "http", ProfilePath, "/debug/pprof/mutex"),
@@ -81,7 +81,7 @@ func TestScrapePool(t *testing.T) {
 				NewTarget(
 					labels.FromStrings("instance", "localhost:9090", "foo", "bar", model.AddressLabel, "localhost:9090", model.MetricNameLabel, pprofProcessCPU, model.SchemeLabel, "http", ProfilePath, "/debug/pprof/profile"),
 					labels.FromStrings("foo", "bar", model.AddressLabel, "localhost:9090", model.MetricNameLabel, pprofProcessCPU, model.SchemeLabel, "http", ProfilePath, "/debug/pprof/profile"),
-					url.Values{"seconds": []string{"14"}},
+					url.Values{"seconds": []string{"17"}},
 				),
 			},
 		},
@@ -103,7 +103,7 @@ func TestScrapePool(t *testing.T) {
 				NewTarget(
 					labels.FromStrings("instance", "localhost:9090", model.AddressLabel, "localhost:9090", model.MetricNameLabel, pprofProcessCPU, model.SchemeLabel, "http", ProfilePath, "/debug/pprof/profile"),
 					labels.FromStrings(model.AddressLabel, "localhost:9090", model.MetricNameLabel, pprofProcessCPU, model.SchemeLabel, "http", ProfilePath, "/debug/pprof/profile"),
-					url.Values{"seconds": []string{"14"}},
+					url.Values{"seconds": []string{"17"}},
 				),
 			},
 		},
@@ -125,7 +125,7 @@ func TestScrapePool(t *testing.T) {
 				NewTarget(
 					labels.FromStrings("instance", "localhost:9090", model.AddressLabel, "localhost:9090", model.MetricNameLabel, pprofProcessCPU, model.SchemeLabel, "http", ProfilePath, "/debug/pprof/profile"),
 					labels.FromStrings("__type__", "foo", model.AddressLabel, "localhost:9090", model.MetricNameLabel, pprofProcessCPU, model.SchemeLabel, "http", ProfilePath, "/debug/pprof/profile"),
-					url.Values{"seconds": []string{"14"}},
+					url.Values{"seconds": []string{"17"}},
 				),
 			},
 		},