+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Main (unreleased)
 
 ### Features
 
-- Grafana Agent Flow: Add `pprof.scrape` and `phlare.write` components to support Profiling.(@cyriltovena)
+- Grafana Agent Flow: Add `pprof.scrape` and `phlare.write` components to support continuous profiling. (@cyriltovena)
 
 - `grafana-agent-operator` supports oauth2 as an authentication method for
   remote_write. (@timo-42)
@@ -887,7 +887,7 @@ v0.19.0 (2021-09-29)
 - Reduced verbosity of tracing autologging by not logging `STATUS_CODE_UNSET`
   status codes. (@mapno)
 
-- Operator: rename Prometheus_CRDs to Metrics_ and Prometheus*fields to
+- Operator: rename Prometheus*CRDs to Metrics* and Prometheus*fields to
   Metrics*. (@rfratto)
 
 - Operator: CRDs are no longer referenced using a hyphen in the name to be